requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
torch==2.0.0
torchvision==0.15.0
numpy==1.21.2
nltk==3.6.2
requests==2.26.0
Pillow==9.0.0
transformers==4.30.2
open_clip_torch==2.20.0